Arrange
I have refactored the Drawing.java, QuardTreeDrawing, 
and add one Interface ArrangeServices and one abstract class Arrange that holds the functionality for Arrange object

+++ JHotDraw/src/main/java/net/n3/nanoxml/StdXMLParser.java
@@ -247,7 +247,7 @@ public class StdXMLParser
                XMLUtil.errorInvalidInput(reader.getSystemID(),
                                          reader.getLineNr(),
                                          "`" + ch + "' (0x"
-                                         + Integer.toHexString((int) ch)
+                                         + Integer.toHexString(ch)
                                          + ')');
          }
       }
